# GetConnection/GetConnection.Api/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Employee.API",
    "version": "v1"
  },
  "paths": {
    "/api/Auth/{validatetoken}": {
      "get": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "validatetoken",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/logOut": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "deviceToken",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/reset_password": {
      "get": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          },
          {
            "in": "query",
            "name": "pswd",
            "type": "string"
          },
          {
            "in": "query",
            "name": "token",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/generate_token": {
      "get": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/check_token": {
      "get": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          },
          {
            "in": "query",
            "name": "reset_token",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth": {
      "get": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          },
          {
            "in": "query",
            "name": "password",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/InsertDeviceToken": {
      "post": {
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "deviceToken",
            "type": "string"
          },
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Hr/ticket/{id}": {
      "get": {
        "tags": [
          "Hr"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Hr/ticket": {
      "post": {
        "tags": [
          "Hr"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddSuportTicketRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Hr": {
      "get": {
        "tags": [
          "Hr"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "sort",
            "type": "string"
          },
          {
            "in": "query",
            "name": "order",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Hr/{id}": {
      "get": {
        "tags": [
          "Hr"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/IssueComplain": {
      "put": {
        "tags": [
          "IssueComplain"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateIssueComplainRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "IssueComplain"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddIssueComplainRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/IssueType/maintenancecategory": {
      "get": {
        "tags": [
          "IssueType"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/IssueType/hrcategory": {
      "get": {
        "tags": [
          "IssueType"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Maintenance/ticket/{id}": {
      "get": {
        "tags": [
          "Maintenance"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Maintenance/ticket": {
      "post": {
        "tags": [
          "Maintenance"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddSuportTicketRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Maintenance": {
      "get": {
        "tags": [
          "Maintenance"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "values",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "sort",
            "type": "string"
          },
          {
            "in": "query",
            "name": "order",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Maintenance/{id}": {
      "get": {
        "tags": [
          "Maintenance"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Organizations": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/InsertOrganizationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/uploadImage": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User": {
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Email",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "EmployeeName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Gender",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "ProfileImage",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "File",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Id",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "formData",
            "name": "Email",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "EmployeeName",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "UserType",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "OrganizationId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "HashToken",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "SaltKey",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "IsActive",
            "type": "boolean"
          },
          {
            "in": "formData",
            "name": "OtpCode",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "File",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "ProfileImage",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/UserID": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "UserID",
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "AddSuportTicketRequest": {
      "type": "object",
      "properties": {
        "issueId": {
          "format": "int64",
          "type": "integer"
        },
        "text": {
          "type": "string"
        },
        "type": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "UpdateIssueComplainRequest": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "deparmentId": {
          "format": "int32",
          "type": "integer"
        },
        "issueTypeID": {
          "format": "int32",
          "type": "integer"
        },
        "employeeId": {
          "format": "int64",
          "type": "integer"
        },
        "statusType": {
          "format": "int32",
          "type": "integer"
        },
        "userDescription": {
          "type": "string"
        },
        "complainDateTime": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "AddIssueComplainRequest": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "deparmentId": {
          "format": "int32",
          "type": "integer"
        },
        "issueTypeID": {
          "format": "int32",
          "type": "integer"
        },
        "employeeId": {
          "format": "int64",
          "type": "integer"
        },
        "statusType": {
          "format": "int32",
          "type": "integer"
        },
        "userDescription": {
          "type": "string"
        },
        "complainDateTime": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "InsertOrganizationRequest": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "organizationName": {
          "type": "string"
        }
      }
    }
  }
}